add some Exception handling

+++ src/main/java/com/example/microserviceforhaulmatic/controller/RolesController.java
@@ -49,25 +49,48 @@ public class RolesController extends Exception {
     }
 
     @DeleteMapping("/{id}")
-    public String deleteEmployerById(@PathVariable String id) {
-        roleService.deleteEmployerById(id);
-        return "deleted";
+    public ResponseEntity<?> deleteEmployerById(@PathVariable String id) {
+        RoleModel roleModel = roleService.getById(id);
+        if (roleModel == null) {
+            return new ResponseEntity<>("NO content for this id", HttpStatus.NO_CONTENT);
+        } else {
+            roleService.deleteEmployerById(id);
+            return new ResponseEntity<>("deleted", HttpStatus.OK);
+        }
+
     }
 
     @PutMapping("/{id}")
-    public RoleModel updateEmployer(@PathVariable String id, @RequestBody RoleDTO roleDTO) {
-        return roleService.updateEmployer(id, roleDTO);
+    public ResponseEntity<?> updateEmployer(@PathVariable String id, @RequestBody RoleDTO roleDTO) {
+        RoleModel roleModel = roleService.getById(id);
+        if (roleModel == null) {
+            return new ResponseEntity<>("NO content for this id", HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity<>(roleService.updateEmployer(id, roleDTO), HttpStatus.OK);
+        }
     }
 
     @ApiOperation(value = "Retrieve a role from the NIC no")
     @GetMapping("nic/{nic}")
-    public RoleModel getByNic(@PathVariable String nic) {
-        return roleService.getByNic(nic);
+    public ResponseEntity<?> getByNic(@PathVariable String nic) {
+        RoleModel roleModel = roleService.getByNic(nic);
+        if (roleModel == null) {
+            return new ResponseEntity<>("No content for thi NIC number", HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity<>(roleModel, HttpStatus.OK);
+        }
     }
 
     @ApiOperation(value = "A list of roles retrieve by the organization and the role type.")
     @GetMapping("/role-type/{roleType}/organization/{organization}")
-    public RoleSpecificDTO getByRoleTypeAndOrganization(@PathVariable String roleType, @PathVariable String organization) {
-        return roleService.getByRoleTypeAndOrganization(roleType, organization);
+    public ResponseEntity<?> getByRoleTypeAndOrganization(@PathVariable String roleType, @PathVariable String organization) throws Exception {
+            RoleSpecificDTO roleSpecificDTO = roleService.getByRoleTypeAndOrganization(roleType, organization);
+            if (roleSpecificDTO == null) {
+                return new ResponseEntity<>("No content for this role and organization", HttpStatus.NO_CONTENT);
+            } else {
+                return new ResponseEntity<>(roleSpecificDTO, HttpStatus.OK);
+            }
+
+
     }
 }
